[#1] (refactor) mobile tree 모듈화.
